mission ok need debug tutor & structure queries
