+++ admin/src/utils/index.js
@@ -197,3 +197,7 @@ export function urlWithScheme(url) {
 export function slugifyFileName(str) {
   return slugify(str, { replacement: "-", remove: /[*+~.()'"!:@]/g });
 }
+
+export function capitalizeFirstLetter(string) {
+  return string[0].toUpperCase() + string.slice(1);
+}
