+++ lib/excelExports.js
@@ -1,0 +1,440 @@
+const translateIndexes = (index) => {
+  switch (index) {
+    case "young":
+      return "volontaires";
+    case "mission":
+      return "Missions";
+    default:
+      return index;
+  }
+};
+const youngExportFields = [
+  {
+    id: "identity",
+    title: "Identité du volontaire",
+    desc: ["Prénom", "Nom", "Sexe", "Cohorte", "Cohorte d'origine"],
+    fields: ["firstName", "lastName", "gender", "cohort", "originalCohort"],
+  },
+  {
+    id: "contact",
+    title: "Contact du volontaire",
+    desc: ["Email", "Téléphone"],
+    fields: ["email", "phone"],
+  },
+  {
+    id: "birth",
+    title: "Date et lieu de naissance du volontaire",
+    desc: [
+      "Date de naissance",
+      "Pays de naissance",
+      "Ville de naissance",
+      "Code postal de naissance",
+    ],
+    fields: ["birthdateAt", "birthCountry", "birthCity", "birthCityZip"],
+  },
+  {
+    id: "address",
+    title: "Lieu de résidence du volontaire",
+    desc: [
+      "Adresse postale",
+      "Code postal",
+      "Ville",
+      "Pays",
+      "Nom de l'hébergeur",
+      "Prénom de l'hébergeur",
+      "Lien avec l'hébergeur",
+      "Adresse - étranger",
+      "Code postal - étranger",
+      "Ville - étranger",
+      "Pays - étranger",
+    ],
+    fields: [
+      "address",
+      "zip",
+      "city",
+      "country",
+      "hostLastName",
+      "hostFirstName",
+      "hostRelationship",
+      "foreignAddress",
+      "foreignZip",
+      "foreignCity",
+      "foreignCountry",
+    ],
+  },
+  {
+    id: "location",
+    title: "Localisation du volontaire",
+    desc: ["Département", "Académie", "Région"],
+    fields: ["department", "academy", "region"],
+  },
+  {
+    id: "schoolSituation",
+    title: "Situation scolaire",
+    desc: [
+      "Niveau",
+      "Type d'établissement",
+      "Nom de l'établissement",
+      "Code postal de l'établissement",
+      "Ville de l'établissement",
+      "Département de l'établissement",
+      "UAI de l'établissement",
+    ],
+    fields: [
+      "situation",
+      "grade",
+      "esSchool.type",
+      "schoolType",
+      "esSchool.fullName",
+      "schoolName",
+      "esSchool.postcode",
+      "schoolZip",
+      "esSchool.city",
+      "schoolCity",
+      "esSchool.department",
+      "schoolDepartment",
+      "esSchool.uai",
+    ],
+  },
+  {
+    id: "situation",
+    title: "Situation particulière",
+    desc: [
+      "Quartier Prioritaire de la ville",
+      "Zone Rurale",
+      "Handicap",
+      "PPS",
+      "PAI",
+      "Aménagement spécifique",
+      "Nature de l'aménagement spécifique",
+      "Aménagement pour mobilité réduite",
+      "Besoin d'être affecté(e) dans le département de résidence",
+      "Allergies ou intolérances alimentaires",
+      "Activité de haut-niveau",
+      "Nature de l'activité de haut-niveau",
+      "Activités de haut niveau nécessitant d'être affecté dans le département de résidence",
+      "Document activité de haut-niveau",
+      "Structure médico-sociale",
+      "Nom de la structure médico-sociale",
+      "Adresse de la structure médico-sociale",
+      "Code postal de la structure médico-sociale",
+      "Ville de la structure médico-sociale",
+    ],
+    fields: [
+      "qpv",
+      "populationDensity",
+      "handicap",
+      "ppsBeneficiary",
+      "paiBeneficiary",
+      "specificAmenagment",
+      "specificAmenagmentType",
+      "reducedMobilityAccess",
+      "handicapInSameDepartment",
+      "allergies",
+      "highSkilledActivity",
+      "highSkilledActivityType",
+      "highSkilledActivityInSameDepartment",
+      "highSkilledActivityProofFiles",
+      "medicosocialStructure",
+      "medicosocialStructureName",
+      "medicosocialStructureAddress",
+      "medicosocialStructureZip",
+      "medicosocialStructureCity",
+    ],
+  },
+  {
+    id: "representative1",
+    title: "Représentant légal 1",
+    desc: [
+      "Statut",
+      "Nom",
+      "Prénom",
+      "Email",
+      "Téléphone",
+      "Adresse",
+      "Code postal",
+      "Ville",
+      "Département et région du représentant légal",
+    ],
+    fields: [
+      "parent1Status",
+      "parent1FirstName",
+      "parent1LastName",
+      "parent1Email",
+      "parent1Phone",
+      "parent1Address",
+      "parent1Zip",
+      "parent1City",
+      "parent1Department",
+      "parent1Region",
+    ],
+  },
+  {
+    id: "representative2",
+    title: "Représentant légal 2",
+    desc: [
+      "Statut",
+      "Nom",
+      "Prénom",
+      "Email",
+      "Téléphone",
+      "Adresse",
+      "Code postal",
+      "Ville",
+      "Département et région du représentant légal",
+    ],
+    fields: [
+      "parent2Status",
+      "parent2FirstName",
+      "parent2LastName",
+      "parent2Email",
+      "parent2Phone",
+      "parent2Address",
+      "parent2Zip",
+      "parent2City",
+      "parent2Department",
+      "parent2Region",
+    ],
+  },
+  {
+    id: "consent",
+    title: "Consentement",
+    desc: ["Consentement des représentants légaux."],
+    fields: ["parentConsentment"],
+  },
+  {
+    id: "status",
+    title: "Statut",
+    desc: [
+      "Statut général",
+      "Statut phase 1",
+      "Statut phase 2",
+      "Statut phase 3",
+      "Date du dernier statut",
+    ],
+    fields: [
+      "status",
+      "phase",
+      "statusPhase1",
+      "statusPhase2",
+      "statusPhase3",
+      "lastStatusAt",
+    ],
+  },
+  {
+    id: "phase1Affectation",
+    title: "Phase 1 - Affectation ",
+    desc: ["ID", "Code", "Nom", "Ville", "Département et région du centre"],
+    fields: [],
+  },
+  {
+    id: "phase1Transport",
+    title: "Phase 1 - Transport",
+    desc: [
+      "Autonomie",
+      "Numéro de bus",
+      "Point de rassemblement",
+      "Dates d'aller et de retour",
+    ],
+    fields: [],
+  },
+  {
+    id: "phase1DocumentStatus",
+    title: "Phase 1 - Statut des documents",
+    desc: [
+      "Droit à l'image",
+      "Autotest PCR",
+      "Règlement intérieur",
+      "Fiche sanitaire",
+    ],
+    fields: [
+      "imageRightFilesStatus",
+      "autoTestPCRFilesStatus",
+      "rulesYoung",
+      "cohesionStayMedicalFileReceived",
+    ],
+  },
+  {
+    id: "phase1DocumentAgreement",
+    title: "Phase 1 - Accords",
+    desc: ["Accords pour droit à l'image et autotests PCR."],
+    fields: ["imageRight", "autoTestPCR"],
+  },
+  {
+    id: "phase1Attendance",
+    title: "Phase 1 - Présence",
+    desc: [
+      "Présence à l'arrivé",
+      "Présence à la JDM",
+      "Date de départ",
+      "Motif de départ",
+    ],
+    fields: [
+      "cohesionStayPresence",
+      "presenceJDM",
+      "departSejourAt",
+      "departSejourMotif",
+    ],
+  },
+  {
+    id: "phase2",
+    title: "Phase 2",
+    desc: [
+      "Domaines MIG 1, MIG 2 et MIG 3",
+      "Projet professionnel",
+      "Période privilégiée",
+      "Choix de périodes",
+      "Mobilité",
+      "Mobilité autour d'un proche",
+      "Information du proche",
+      "Mode de transport",
+      "Format de mission",
+      "Engagement hors SNU",
+      "Souhait MIG",
+    ],
+    fields: [
+      "domains",
+      "professionnalProject",
+      "professionnalProjectPrecision",
+      "period",
+      "periodRanking",
+      "mobilityNearSchool",
+      "mobilityNearHome",
+      "mobilityNearRelative",
+      "mobilityNearRelativeName",
+      "mobilityNearRelativeAddress",
+      "mobilityNearRelativeZip",
+      "mobilityNearRelativeCity",
+      "mobilityTransport",
+      "missionFormat",
+      "engaged",
+      "engagedDescription",
+      "desiredLocation",
+    ],
+  },
+  {
+    id: "accountDetails",
+    title: "Compte",
+    desc: ["Dates de création, d'édition et de dernière connexion"],
+    fields: ["createdAt", "updatedAt", "lastLoginAt"],
+  },
+  {
+    id: "desistement",
+    title: "Désistement",
+    desc: ["Raison du désistement", "Message de désistement"],
+    fields: ["withdrawnReason", "withdrawnMessage"],
+  },
+];
+
+const missionExportFields = [
+  {
+    id: "missionInfo",
+    title: "Informations sur la mission",
+    desc: [
+      "ID",
+      "Titre de la mission",
+      "Date du début",
+      "Date de fin",
+      "Nombre de volontaires recherchés",
+      "Places restantes sur la mission",
+      "Visibilité de la mission",
+      "Source de la mission",
+    ],
+    fields: [
+      "_id",
+      "name",
+      "startAt",
+      "endAt",
+      "placesTotal",
+      "placesLeft",
+      "visibility",
+      "isJvaMission",
+    ],
+  },
+  {
+    id: "status",
+    title: "Statut de la mission",
+    desc: ["Statut de la mission", "Créée lé", "Mise à jour le"],
+    fields: ["status", "createdAt", "updatedAt"],
+  },
+  {
+    id: "missionType",
+    title: "Type de mission",
+    desc: [
+      "Domaine principal de la mission",
+      "Domaine(s) secondaire(s) de la mission",
+      "Préparation militaire",
+    ],
+    fields: ["mainDomain", "format", "isMilitaryPreparation"],
+  },
+  {
+    id: "missionDetails",
+    title: "Détails de la mission",
+    desc: [
+      "Objectifs de la mission",
+      "Actions concrètes",
+      "Contraintes",
+      "Durée",
+      "Fréquence estimée",
+      "Période de réalisation",
+    ],
+    fields: [
+      "description",
+      "actions",
+      "contraintes",
+      "duration",
+      "frequence",
+      "period",
+    ],
+  },
+  {
+    id: "tutor",
+    title: "Tuteur de la mission",
+    desc: [
+      "Id du tuteur",
+      "Nom du tuteur",
+      "Prénom du tuteur",
+      "Email du tuteur",
+      "Portable du tuteur",
+      "Téléphone du tuteur",
+    ],
+    fields: ["tutorID"],
+  },
+  {
+    id: "location",
+    title: "Localisation de la mission",
+    desc: ["Adresse", "Code postal", "Ville", "Département", "Région"],
+    fields: ["address", "zip", "city", "department", "region"],
+  },
+  {
+    id: "structureInfo",
+    title: "Informations sur la structure",
+    desc: [
+      "Id de la structure",
+      "Nom de la structure",
+      "Statut juridique de la structure",
+      "Type(s) de structure",
+      "Sous-type de structure",
+      "Présentation de la structure",
+    ],
+    fields: ["structureID"],
+  },
+  {
+    id: "structureLocation",
+    title: "Localisation de la structure",
+    desc: [
+      "Adresse de la structure",
+      "Code postal de la structure",
+      "Ville de la structure",
+      "Département de la structure",
+      "Région de la structure",
+    ],
+    fields: ["structureID"],
+  },
+];
+
+module.exports = {
+  translateIndexes,
+  youngExportFields,
+  missionExportFields,
+};
